generator/requirements.txt
psycopg[binary]==3.2.10
python-dotenv==1.0.1
Faker==26.0.0